Allow Goto dialog to support line number text to have non numeric suffixes
This patch enables the line number field in the Goto dialog to accept text with
non numeric characters suffixes. For example when you do grep/ripgrep the output
contains something like below.

```
    D:\repos\notepad2\src>rg LINENUM
    Notepad2.rc
    404:        MENUITEM "Line &Numbers\tCtrl+Shift+N",     IDM_VIEW_LINENUMBERS
    745:    "N",            IDM_VIEW_LINENUMBERS,   VIRTKEY, SHIFT, CONTROL, NOINVERT
    1055: IDD_LINENUM DIALOGEX 0, 0, 166, 70
    1061:    EDITTEXT        IDC_LINENUM,7,18,90,14,ES_AUTOHSCROLL
    1535:    IDD_LINENUM, DIALOG
```

It is so much easier to copy the line number by double clicking on the line number
in the CMD and paste it in the Goto dialog. Without this fix we cannot directly
use the copied text as it includes the semi-colon(':')