Enable "Open ANSI (unknown encoding) file in UTF-8 mode" by default, issue #386.
Change the implementation to use system ANSI code page to decode
the opening file (which is not valid UTF-8), if decoding succeeded
without data loss then convert the decoded UTF-16 result into UTF-8.

This diff from previous behavior, which always use UTF-8 to decode
the opening file when the option is enabled and the file can not be
detected as UTF-16.