Statusbar Enhancement (#503)
* Statusbar Enhancement

Improved the statusbar with panels that improve the readability of document information (Ln, Col, Ch, etc.).
The width of the panels dynamically updates according to the size of the window and the text contained.
The Ln and Fnd panels double-click to open the 'go to' and 'find and replace' windows respectively. All location files have been updated with the new variables.

* Optimisation of nomenclatures and styles

Updated the files to be more in keeping with the overall style of the code. Inserted small 'for' loop to simplify code in Notepad.c.