Fix document size label and separator width
After remaking the statusbar code, the document size indicator no longer works. In some cases only when opening a file it is shown but then remains unchanged even when the document is changed.
By removing an 'if' on updateMask everything works fine but it is a crude fix.
The spacing of the separators in the statusbar seems excessive to me (which is why I changed it) but it is a personal taste so it might as well stay as it is.